Competitive offer generator

+++ src/main/java/nure/abittour/dto/ApplicationDto.java
@@ -2,11 +2,13 @@ package nure.abittour.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ApplicationDto {
     private Long id;
     private Long studentId;
     private Long competitiveOfferId;
-    private Double totalScore;
+    private BigDecimal totalScore;
     private Integer priority;
 }
